Move Choreo model within client package
