+++ choreo-extension-native/src/main/java/io/ballerina/observe/choreo/client/ChoreoClient.java
@@ -28,9 +28,9 @@ import io.ballerina.observe.choreo.gen.TelemetryGrpc;
 import io.ballerina.observe.choreo.gen.TelemetryOuterClass;
 import io.ballerina.observe.choreo.logging.LogFactory;
 import io.ballerina.observe.choreo.logging.Logger;
-import io.ballerina.observe.choreo.model.ChoreoMetric;
-import io.ballerina.observe.choreo.model.ChoreoTraceSpan;
-import io.ballerina.observe.choreo.model.SpanEvent;
+import io.ballerina.observe.choreo.client.model.ChoreoMetric;
+import io.ballerina.observe.choreo.client.model.ChoreoTraceSpan;
+import io.ballerina.observe.choreo.client.model.SpanEvent;
 import io.grpc.ManagedChannel;
 import io.grpc.ManagedChannelBuilder;
 import io.grpc.StatusRuntimeException;
