+++ choreo-extension-native/src/main/java/io/ballerina/observe/choreo/client/model/SpanEvent.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package io.ballerina.observe.choreo.model;
+package io.ballerina.observe.choreo.client.model;
 
 /**
  * Represents a Choreo Trace Span Event.
